+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/blockentity/RecipeRegistryBlockEntity.kt
@@ -5,7 +5,6 @@ import net.minecraft.core.Direction
 import net.minecraft.world.level.block.state.BlockState
 import site.siredvin.peripheralium.common.blockentities.PeripheralBlockEntity
 import site.siredvin.peripheralworks.common.setup.BlockEntityTypes
-import site.siredvin.peripheralworks.computercraft.peripherals.RealityForgerPeripheral
 import site.siredvin.peripheralworks.computercraft.peripherals.RecipeRegistryPeripheral
 
 class RecipeRegistryBlockEntity(blockPos: BlockPos, blockState: BlockState) :