+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/data/ModLootTableProvider.kt
@@ -31,5 +31,6 @@ object ModLootTableProvider {
         LootTableHelper.dropSelf(consumer, Blocks.PERIPHERAL_PROXY)
         LootTableHelper.dropSelf(consumer, Blocks.REALITY_FORGER)
         LootTableHelper.dropSelf(consumer, Blocks.RECIPE_REGISTRY)
+        LootTableHelper.dropSelf(consumer, Blocks.INFORMATIVE_REGISTRY)
     }
 }