+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/subsystem/recipe/NBTCheckMode.kt
@@ -7,20 +7,26 @@ enum class NBTCheckMode {
     FULL, SUBSET, SUPERSET, NONE;
 
     fun itemStackEquals(result: ItemStack, targetingResult: ItemStack): Boolean {
-        if (result.item != targetingResult.item)
-            return false;
-        if (this == NONE)
-            return true;
-        val targetingNBT = targetingResult.tag;
-        val resultNBT = result.tag;
-        if (targetingNBT == null)
-            return resultNBT == null || this == SUBSET;
-        if (resultNBT == null)
-            return this == SUPERSET;
-        if (this == FULL)
-            return resultNBT == targetingNBT;
-        if (this == SUBSET)
-            return NBTUtil.isSubSet(resultNBT, targetingNBT);
-        return NBTUtil.isSubSet(targetingNBT, resultNBT);
+        if (result.item != targetingResult.item) {
+            return false
+        }
+        if (this == NONE) {
+            return true
+        }
+        val targetingNBT = targetingResult.tag
+        val resultNBT = result.tag
+        if (targetingNBT == null) {
+            return resultNBT == null || this == SUBSET
+        }
+        if (resultNBT == null) {
+            return this == SUPERSET
+        }
+        if (this == FULL) {
+            return resultNBT == targetingNBT
+        }
+        if (this == SUBSET) {
+            return NBTUtil.isSubSet(resultNBT, targetingNBT)
+        }
+        return NBTUtil.isSubSet(targetingNBT, resultNBT)
     }
 }