+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `additional lanterns` integration for fabric
 - Reality forger
 - Recipe registry
+- Informative registry
 
 ## [0.5.3] - 2023-07-01